29_ Part 28. 테스트 Ch 5. Genre Module Unit Test

+++ src/director/director.service.ts
@@ -27,11 +27,12 @@ export class DirectorService {
   }
 
   async update(id: number, updateDirectorDto: UpdateDirectorDto) {
-    const director = await this.directorRepository.exists({
+    const director = await this.directorRepository.findOne({
       where: { id },
     });
-    if (!director)
+    if (!director) {
       throw new NotFoundException('존재하지 않는 ID의 영화입니다.');
+    }
 
     await this.directorRepository.update(
       {
@@ -44,7 +45,15 @@ export class DirectorService {
     });
   }
 
-  remove(id: number) {
-    return this.directorRepository.delete(id);
+  async remove(id: number) {
+    const director = await this.directorRepository.findOne({
+      where: { id },
+    });
+    if (!director) {
+      throw new NotFoundException('존재하지 않는 ID의 영화입니다.');
+    }
+    await this.directorRepository.delete(id);
+
+    return id;
   }
 }
